+++ src/index.ts
@@ -1,15 +1,6 @@
 /* Everything exported here is considered public API and is documented by typedoc.
  * DO NOT DELETE THIS FILE
  */
-
-export {
-  Status,
-  LegacySudoSiteReputationClient,
-  LegacySudoSiteReputationClientProps,
-  LegacySiteReputation,
-  lastUpdatePerformedAtStorageKey,
-  malwareRulesetStorageKey,
-} from './legacy-sudo-site-reputation-client'
 export {
   ReputationStatus,
   SiteReputation,
@@ -17,14 +8,6 @@ export {
   SudoSiteReputationClientProps,
 } from './sudo-site-reputation-client'
 export { ApiClient } from './apiClient'
-export { Config, IotsConfig } from './config'
 export * as Entitlements from './entitlements'
 export { StorageProvider } from './storage-provider'
-export { RulesetType } from './ruleset-type'
-export { ReputationDataNotPresentError } from './errors'
-export * from './ruleset-provider'
-export {
-  DefaultRulesetProvider,
-  DefaultRulesetProviderProps,
-} from './default-ruleset-provider'
 export { MemoryStorageProvider } from './memory-storage-provider'