+++ src/ruleset-type.ts
@@ -1,9 +1,0 @@
-/**
- * An enum describing types of rulesets.
- */
-export enum RulesetType {
-  /** A ruleset that blocks malicious domains */
-  MALICIOUSDOMAIN = 'MALICIOUSDOMAIN',
-  MALWARE = 'MALWARE',
-  PHISHING = 'PHISHING',
-}