+++ src/sudo-site-reputation-client.spec.ts
@@ -8,6 +8,7 @@ const reputationCache = new IDBStorageProvider()
 const apiClient = {
   getSiteReputation: jest.fn().mockResolvedValue({
     reputationStatus: ReputationStatus.Malicious,
+    categories: [],
   }),
 } as any
 
@@ -19,29 +20,38 @@ describe('SudoSiteReputationClient', () => {
   })
 
   it('get a known malicious site', async () => {
+    apiClient.getSiteReputation.mockResolvedValue({
+      reputationStatus: ReputationStatus.Malicious,
+      categories: ['MEAN_COFFEE_DRINKERS'],
+    })
     const result = await client.getSiteReputation('someurl.com')
     expect(result).toEqual({
       reputationStatus: ReputationStatus.Malicious,
+      categories: ['MEAN_COFFEE_DRINKERS'],
     })
   })
 
   it('get a known safe site', async () => {
     apiClient.getSiteReputation.mockResolvedValue({
       reputationStatus: ReputationStatus.Notmalicious,
+      categories: [],
     })
     const result = await client.getSiteReputation('someurl.com')
     expect(result).toEqual({
       reputationStatus: ReputationStatus.Notmalicious,
+      categories: [],
     })
   })
 
   it('get a unknown site', async () => {
     apiClient.getSiteReputation.mockResolvedValue({
       reputationStatus: ReputationStatus.Unknown,
+      categories: [],
     })
     const result = await client.getSiteReputation('someurl.com')
     expect(result).toEqual({
       reputationStatus: ReputationStatus.Unknown,
+      categories: [],
     })
   })
 
@@ -52,19 +62,31 @@ describe('SudoSiteReputationClient', () => {
     const unknownResult = await client.getSiteReputation(
       'http://someunknownsite.com',
     )
-    expect(unknownResult.reputationStatus).toBe(ReputationStatus.Unknown)
+    expect(unknownResult).toEqual({
+      reputationStatus: ReputationStatus.Unknown,
+      categories: [],
+    })
     await expect(
       reputationCache.getItem('http://someunknownsite.com'),
-    ).resolves.toEqual(ReputationStatus.Unknown)
+    ).resolves.toEqual(
+      JSON.stringify({
+        reputationStatus: ReputationStatus.Unknown,
+        categories: [],
+      }),
+    )
     await reputationCache.setItem(
       'http://somenonmalicioussite.com',
-      ReputationStatus.Notmalicious,
+      JSON.stringify({
+        reputationStatus: ReputationStatus.Notmalicious,
+        categories: [],
+      }),
     )
     const nonMaliciousResult = await client.getSiteReputation(
       'http://somenonmalicioussite.com',
     )
-    expect(nonMaliciousResult.reputationStatus).toBe(
-      ReputationStatus.Notmalicious,
-    )
+    expect(nonMaliciousResult).toEqual({
+      reputationStatus: ReputationStatus.Notmalicious,
+      categories: [],
+    })
   })
 })