+++ tests/system/test-registration.ts
@@ -7,6 +7,12 @@ import {
   KeyDataKeyType,
   SudoKeyManager,
 } from '@sudoplatform/sudo-common'
+import { Entitlement } from '@sudoplatform/sudo-entitlements'
+import {
+  DefaultSudoEntitlementsAdminClient,
+  ExternalUserEntitlements,
+  SudoEntitlementsAdminClient,
+} from '@sudoplatform/sudo-entitlements-admin'
 import {
   DefaultSudoUserClient,
   SudoUserClient,
@@ -14,6 +20,7 @@ import {
 } from '@sudoplatform/sudo-user'
 import { WebSudoCryptoProvider } from '@sudoplatform/sudo-web-crypto-provider'
 import * as dotenv from 'dotenv'
+import { sortBy, uniqBy } from 'lodash'
 
 import { requireEnv } from '../../utils/require-env'
 import { logger } from './logger'
@@ -26,7 +33,9 @@ interface Env {
   SDK_CONFIG: string
 }
 
-function loadEnv(): Env {
+export const SR_ENTITLEMENT_NAME = 'sudoplatform.sr.srUserEntitled'
+
+export function loadEnv(): Env {
   const env: Partial<Env> = {}
   Object.keys(process.env)
     .sort()
@@ -89,9 +98,14 @@ const testAuthProvider = new TESTAuthenticationProvider(
   env.REGISTER_KEY_ID,
 )
 
+const adminEnv = requireEnv({
+  ADMIN_API_KEY: { type: 'string' },
+})
+
 export async function registerUser(): Promise<{
   keyManager: SudoKeyManager
   userClient: SudoUserClient
+  entitlementAdminClient: SudoEntitlementsAdminClient
 }> {
   const cryptoProvider = new WebSudoCryptoProvider('ns', 'sr-service')
   const keyManager = new DefaultSudoKeyManager(cryptoProvider)
@@ -99,11 +113,21 @@ export async function registerUser(): Promise<{
     sudoKeyManager: keyManager,
     logger,
   })
+  const entitlementAdminClient = new DefaultSudoEntitlementsAdminClient(
+    adminEnv.ADMIN_API_KEY,
+  )
 
+  // Register
   await userClient.registerWithAuthenticationProvider(testAuthProvider)
   await userClient.signInWithKey()
 
-  return { userClient, keyManager }
+  // Add entitlements
+  const externalId = (await userClient.getUserName()) ?? ''
+  await entitlementAdminClient.applyEntitlementsToUser(externalId, [
+    { name: SR_ENTITLEMENT_NAME, value: 1 },
+  ])
+
+  return { userClient, keyManager, entitlementAdminClient }
 }
 
 export async function invalidateAuthTokens(
@@ -135,4 +159,55 @@ export async function invalidateAuthTokens(
         throw new Error(`Unknown KeyDataKeyType: ${value.type}`)
     }
   })
+}
+
+export async function removeEntitlementsByName(
+  externalId: string,
+  entitlementNamesToRemove: string[],
+  entitlementAdminClient: SudoEntitlementsAdminClient,
+): Promise<void> {
+  const { entitlements } = await entitlementAdminClient.getEntitlementsForUser(
+    externalId,
+  )
+  const updatedEntitlements = entitlements.entitlements.filter(
+    (userEntitlement) =>
+      !entitlementNamesToRemove.some(
+        (entitlementNameToRemove) =>
+          entitlementNameToRemove === userEntitlement.name,
+      ),
+  )
+  await entitlementAdminClient.applyEntitlementsToUser(
+    externalId,
+    updatedEntitlements,
+  )
+}
+
+export async function updateEntitlements(
+  externalId: string,
+  entitlementsToUpdate: Entitlement[],
+  entitlementAdminClient: SudoEntitlementsAdminClient,
+): Promise<ExternalUserEntitlements> {
+  const userEntitlements = await entitlementAdminClient.getEntitlementsForUser(
+    externalId,
+  )
+
+  const existingEntitlements = (
+    userEntitlements.entitlements.entitlements ?? []
+  ).map((e) => ({
+    name: e.name,
+    value: e.value,
+  }))
+
+  const entitlementsToApply = sortBy(
+    uniqBy(
+      [...entitlementsToUpdate, ...existingEntitlements],
+      (entitlement) => entitlement.name,
+    ),
+    'name',
+  )
+
+  return entitlementAdminClient.applyEntitlementsToUser(
+    externalId,
+    entitlementsToApply,
+  )
 }